requirements.txt
git+https://github.com/coinbase/coinbase-advanced-trade-python.git
python-dotenv
pandas
